Refactor API
This refactor uses a more useful abstraction. In this abstraction, only analytical logpdf corrections are computed simultaneously with the constrained value if available. Autodiff handles the adjoints well enough that we don't gain anything by hand-coding them, and we no longer compute intermediate quantities for adjoints when gradients are not requested. To simplify the code-base, we remove default workarounds for type inference and leave it to the user to ensure that their special case is type inferred.